Optimize and refactor db access functions
