+++ lib/commands.js
@@ -1,10 +1,10 @@
-const { readPasswords, writePasswords } = require('./passwords');
+const { readPasswords, writePasswords } = require('./dbaccess');
 const { encrypt, decrypt } = require('./crypto');
 
 function get(key) {
   try {
     const passwords = readPasswords();
-    const encryptedPassword = passwords.passwords[key];
+    const encryptedPassword = passwords[key];
     const password = decrypt(encryptedPassword);
     console.log(`${key}: "${password}"`);
   } catch (error) {
@@ -16,7 +16,7 @@ function set(key, value) {
   const encryptedValue = encrypt(value);
   try {
     const passwords = readPasswords();
-    passwords.passwords[key] = encryptedValue;
+    passwords[key] = encryptedValue;
     writePasswords(passwords);
   } catch (error) {
     console.error(error);
@@ -26,7 +26,7 @@ function set(key, value) {
 function unset(key) {
   try {
     const passwords = readPasswords();
-    delete passwords.passwords[key];
+    delete passwords[key];
     writePasswords(passwords);
   } catch (error) {
     console.error(error);
