+++ lib/crypto.js
@@ -1,5 +1,5 @@
 const crypto = require('crypto');
-const { readMasterPassword } = require('./passwords');
+const { readMasterPassword } = require('./dbaccess');
 
 let resizedIV = Buffer.allocUnsafe(16);
 const iv = crypto.createHash('sha256').update('myHashedIV').digest();
